feat: add configuration options for handling vcfs
though the flags are created, they are not received anywhere. the old filter files flag is modified
to specify that it's for filtered info files. I'm not sure I like this change, and may revert it if
I think that ultimately a second-pass filter should imply all inputs are always filtered.

re #2